upgrade to current x25519-dalek
The latest x25519-dalek removes access to the internal scalar of secret-key
objects (no EphemeralSecret.as_bytes), so instead we must use the low-level
x25519() function on a slice of bytes. These bytes are unclamped, whereas
previously the private keys we printed were clamped, so we're depending upon
Wireguard to clamp the private-key scalars it reads from the config file.