allow match to *start* within 10 chars, doesn't have to finish
This means the speedup is 10x regardless of the length of the prefix.
Previously, we only looked at the first 10 chars, so a 5-char match that
began at offset 8 would be ignored, and the speedup would be 5x for a 5-char
prefix, 4x for a 6-char prefix, and more generally (10-n)x for a n-char
prefix.